requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0